Use range.contains and idiomatic let-chains
Replace explicit comparison checks with range.contains, adopt let-chain
conditionals for concise option checking, remove unnecessary casts and
redundant return, and tidy up a few arithmetic expressions for clarity

+++ crates/large-track-viewer/src/app/mod.rs
@@ -259,8 +259,8 @@ impl LargeTrackViewerApp {
             let center_lat = (min_lat + max_lat) / 2.0;
             let center_lon = (min_lon + max_lon) / 2.0;
 
-            let lat_span = (max_lat as f64 - min_lat as f64).abs();
-            let lon_span = (max_lon as f64 - min_lon as f64).abs();
+            let lat_span = (max_lat - min_lat).abs();
+            let lon_span = (max_lon - min_lon).abs();
             let max_span = lat_span.max(lon_span);
 
             let zoom = if max_span > 0.0 {
